refactor: change App refference
